student show page

+++ app/controllers/guardians_controller.rb
@@ -9,11 +9,9 @@ class GuardiansController < ApplicationController
   end
 
   def create
-    p "*******************************"
     @guardian = Guardian.new(
                               first_name: params[:first_name],
                               last_name: params[:last_name],
-                              # email: params[:email], migrated out of schema due to devise--may be ok written as is
                               home_phone: params[:home_phone],
                               cell_phone: params[:cell_phone],
                               address: params[:address],
@@ -22,9 +20,10 @@ class GuardiansController < ApplicationController
                               zip: params[:zip],
                               country: params[:country]
                               )
-    @guardian.save 
-    flash[:success] = "Account Created"
-    redirect_to "/guardians#{current_guardian.id}"
+      @guardian.save 
+      flash[:success] = "Account Created"
+      redirect_to "/guardians#{current_guardian.id}"
+
 
   end
 
@@ -42,22 +41,31 @@ class GuardiansController < ApplicationController
     @guardian.assign_attributes(
                               first_name: params[:first_name],
                               last_name: params[:last_name],
-                              # email: params[:email], migrated out of schema due to devise--may be ok written as is
                               home_phone: params[:home_phone],
                               cell_phone: params[:cell_phone],
                               address: params[:address],
                               city: params[:city],
                               state: params[:state],
-                              zip: params[:zip]
+                              zip: params[:zip],
+                              country: params[:country]
                               )
-    @guardian.save
-    flash[:success] = "Account Updated"
-    redirect_to "/guardians/#{current_guardian.id}"
+    if @guardian.save
+      flash[:success] = "Account Updated"
+      redirect_to "/guardians/#{current_guardian.id}"
+    else 
+      flash[:warning] = "Error with Information, try again"
+      redirect_to "/guardians/#{current_guardian.id}/edit"
+    end
   end
 
   def destroy
-    @guardian = Guardian.find(params[:id])
-    guardian.destroy
+    if @guardian = Guardian.find(params[:id])
+       guardian.destroy
+       flash[:sucess] = "Account Deleted"
+    else
+      flash[:warning] = "Error - try again"
+      redirect_to "/guardians/#{current_guardian.id}"
+    end 
   end
 
 end
